Update appearance.md
The example for "Change the Order of the Time and Calendar Buttons" only works if there is one DateTimePicker on the page. If there are two it will put the first button from the first control next to the last one in the second control.

My proposal will loop through them all and fix them properly no matter how many exist.